Correct .gitignore syntax and remove compiled bootstrap.json file at root.
Prior to this commit, the git repo was omitting files from the
generated directories (build/, phonegap/, node_modules/) but not their
respective subdirectories. This commit corrects .gitignore syntax to
ensure these subdirectories do not get re-introduced to the project.

This commit also removes the bootstrap.json file at the root of the
project as it is also generated on 'sencha web start' and 'sencha app
build'.