Fixing map centering on device
I was hindered by the lack of wifi in our work environment. This group of
changes works. I moved the scroll to function into the map view init
function, because otherwise it was not being run when moving between the
home and list screens. This occured, because the gMap value was
instaniated after the panTo function in the controller.