Update map markers
I updated the map markers so that they would display an infowindow when
clicked on that shows the store name at that location. I also made it so
that only one infowindow shows at a time. There is also commented out code
in the sass formating folder that we can use to style and change the look
of the infowindows.Also removed the previous search's map markers so that
they didn't persist through different queries.