Add a button to the infowindow that has navigation
I added a button to the info window that lets the user navigate to the
detail page and view the selected vendor's items. I still need to update
the info and header sections of the detail window when navigating there
from a map marker. To do this I will need to set the toolbar in the
controller and set the detail view data to the selected vendors data.