Add list item selection on infowindow button click
When a button is clicked in a google maps marker infowindow, the
coresponding list item for that vendor is highlighted. We could face
problems in the future when the list is populated with products. We will
deal with that advanced functionality when we get to loading products into
the list.