Add renamed store and model to app.js requirements arrays
Prior to this commit, the list page was unable to properly handle list
elements because those resources had been renamed and were therefore
not available to Ext. This commit just reintroduces the model and
store for handling list elements (now named "VendorInventories" and
"VendorInventory" respectively) for handling list items to the app.js
requirement arrays.