Change map marker colors when item is highlighted
To do this I broke out the functions to add a marker to the map from the addMapMarkers function and made it a new function called addAMapMarker. In the
onViewLpageListHighlightCommand I checked which type of list we had,
vendors or products. If vendors list, when a list item was highlighted we
made that map marker blue and popped up its infowindow. If products list,
when a list item was highlighted we made the map markers of vendors who
carried that specific item blue. We called the
onViewLpageListHighlightCommand in the onViewLpageListItemCommand so that
when a user navigates back to the list page from a detail page, the
correct list item and is selected and the marker(s) are blue. In
order for a user to easily see the locations of vendors, on a single tap
we will bring the markers to the front and color them blue.