Make list dislplay data depending checkboxes
I was able to make it so that a user could either see vendors or products
in the list view by selecting the corresponding checkbox. Also only one
checkbox could be used at a time. The checkboxes set the store that list
page uses to populate its list. In order to do this I also had to break
out some from the onChooseProduct function in order to make the
populatePstore function so that we could populate the productList store
with the current data that we were searching for. I broke out this new
function and the numberOfVendors function so that the code would be "DRY".