Fix getAt(i) for Second Test Case Expect
In a prior commit, in 'loads multple datas from an api' function,  getAt(1)
for the second test case expect was accidentally changed to getAt(0).  Because
there are two tests cases in this function, the second test case expect should
use getAt(1). This commit reverts getAt(0) back to getAt(1) for the second test
case expect.