Make only specific labels dissappear
Changed all labels to just points of interest adminstrative and landscape.
This commit will make it so that buses and road names don't disappear from the map as well.
This works on my machine, but the user may have to refresh cache and rebuild the project after applying this change.