Fix list item selection on infowindow select
I made it so that when an infowindow was selected, the correct list item
was highlighted when the user navigated back to the list page. If a list of vendors was showing, the vendor was highlighted,
and if a list of products was showing, then the first product that the
vendor sold would be highlighted.