Add callback function to Locations and Products store autoload settings
Prior to this commit, the selectfields on the homepage of the app were
unable to display "empty" options such as "Please choose a product"
because the api did not include this option. This commit adds a
callback function to the stores' autoload settings to insert the
"empty" field option. The text in this option may be changed at
anytime. The valueField option in the home screen was updated to allow
the selectfields to index properly.