Abstract onChooseLocation and onChooseProduct filter functionality to helper function
This commit introduces the filterVendorStore, matchesLocation and
matchesProduct helper functions to break out the filtering
functionality common to the onChooseLocation and onChooseProduct
controller functions. The helper function allows the controller to
'and' the search criteria from the location and product fields on the
home screen.

This commit does not resolve the issue of page initialization. The
home screen still loads with incorrect options for the location and
product select fields.