Add "location" field to the Location model to capture API index data.
Prior to this commit, the location selectField was not properly
indicating which field was already selected. It would highlight
whichever element was first in the store regardless of which field the
user had selected. This behavior is a known bug in the Sencha
framework that occurs when a selectField does not have a default
index or "valueField". This commit leverages the index provided by the
whats-fresh-api to provide this default index (called 'location').