Add listener to vendor store to repaint view when API data is properly loaded.
Prior to this commit, the first time a user would attempt to load the
home screen or press the "Go" button without filtering by any
criteria, the search results would appear empty. This was due to
asynchronous loading in the vendor store. This commit introduces a
listener to the vendor store that repaints the search summary when the
store is loaded from the API.