add PointOfInterest store
This commit adds a store for the PointOfInterest model.

This commit resolves the failing test on:
WorkingWaterfronts.store.PointOfInterest

As of this commit, the project will not run due to the stores and
models removed in c9c86b1. These will need to be sync’d in the app’s
root.