redesign home view, add home controller
This commit is an intermediate step for creating Home’s controller.
Before this commit, WWM did not have a controller for the home view.
After this commit, a partially completed controller and util classes
exist.

This commit is not whole due to lack of unit tests.

I discovered that unit tests cannot be done from the browser with our
project design, as files existing higher than SpecRunner.html are
considered cross-domain by Ajax. Either SpecRunner.html will need to
moved to the project root, or we need to use programmatic testing with
a tool like Karma, examples for which are provided by PivotalLabs.