recreate listview as maplist
This commit contains work towards making the map and list view work.

A new MapList view was created, containing functionality similar to old
ListView. Map-specific functions needs to be moved to the map view, and
the MapList needs to place markers for each point of interest filtered,
done within refresh event handler.

The PointOfInterest store is gone, and the working and dummy versions
are available in its place. Tests fail with the missing file.