pare down list view
Before this commit, old list view functionality was disconnected from
the Home view. After this commit, a new MapList view and controller
displays the contents of the PointOfInterest store. The map and list
stay in sync with their current selection. Some marker and centering
functionality has been moved to the Map.js view.

The next step is to create a detail view for points of interest.