Add 'waiting double tap' to MapList view
Before this commit, CSS styles displayed drop-downs weird, and the
MapList did not support clicking on a highlighted element to open its
POI detail view.

After this commit, the CSS styles have been repaired, and clicking on a
POI that is already selected will take you to the detail view, in
addition to double tapping and clicking the button in the Google Maps
InfoWindow. In addition, the map will now pan to selected items.