Overhaul project directory for Phonegap building
Before this commit, the project was not tuned for automatic builds, and
contained experiments in configuring Cordova and Phonegap.

After this commit, the project has been reinitialized. Only required
configuration files remain, the rest have been pruned. Some resources
moved into the right resource location.

It is important to note that the settings in .sencha have been changed.
The project was build with a newer version of Sencha Cmd, so developers
might need to update their environment.