Update phonegap/config.xml to properly display .9.png splash screens
Prior to this commit, phonegap was not configured to properly display
splash screen files that were properly copied into the project
directory. This commit addresses this problem by updating the
configuration to omit ".9.png" from the names of the linked files.

The need for this change was in no way represented by Phonegap's
documentation.