Remove all references to Location selector in view/Home.js
Prior to this commit, the app was loading and maintaining the location
store, which is not used at all in the deployment version of Working
Waterfronts because we no longer allow users to select by City. This
commit eliminates all models and stores for Location information and
removes all references to the Location selector in view/Home.js.