Update API references to use production API location
Prior to this commit, references to the Working Waterfronts API in the
store and the view used a "hosturl" global variable. This commit
updates this reference to use a util.API Sencha class (unifying style
between this project and Oregon's Catch).

This commit also points the new util.API sencha class at the
production Working Waterfronts API location and updates the cordova
whitelist <access> tag in phonegap/config.xml accordingly.