Change map callout <a> tag to <button>
Prior to this commit, users clicking on the map callout would cause
the android version of the app to open a new, empty tab in the
browser. This commit avoids this escape behavior by changing the <a>
tag to a <button> that has been styled to appear as a link.